feat: init tui (#46)
Adds a TUI with two panes:
1. Wallets
2. Dashboard

Wallets pane lists all the keystores in the foundry keystore
`~/.foundry/keystores`. User is able to navigate the list of keystores
and select any keystore as "active" wallet which will prompt for a
password if the keystore is not already unlocked.

Active pane is highlighted with blue borders. The events are forwarded
to the relevant active pane/element.

In future, the dashboard pane will show all pending requests(connect to
website, transaction signature, eip712 signature, eip7702 authorization
signature, etc.).

Additional signer types will be integrated later.